+++ packages/easy-email-extensions/src/KiaLayout/index.module.scss
@@ -65,11 +65,8 @@
 .blockItemWrapperSmall {
   width: 100px;
   display: flex;
-  //margin: 0.5em 0;
   padding-right: 0.3em;
   border-radius: 15px;
-  //background-color:salmon;
-  //font-size:x-small
 }
 
 .blockItem {
@@ -87,10 +84,3 @@
 
 }
 
-.blockHeading {
-  //text-align: center;
-  padding: 3px 12px;
-  min-width: 60px;
-  font-weight: bold;
-
-}
