+++ packages/easy-email-extensions/src/KiaLayout/leftnav/BlockItem.tsx
@@ -1,0 +1,24 @@
+import { Help } from '@extensions/AttributePanel/components/UI/Help';
+import React from 'react';
+import styles from '../index.module.scss';
+import { IBlockCategoryItem } from './model/BlockCategory';
+
+export const BlockItem = ({ block }: { block: IBlockCategoryItem }) => {
+  return (
+    <div
+      className={
+        block.subItem ? styles.blockItemWrapperSmall : styles.blockItemWrapper
+      }
+    >
+      <div
+        className={block.subItem ? styles.blockItemSmall : styles.blockItem}
+        style={{ flexGrow: 1 }}
+      >
+        {block.title}{' '}
+      </div>
+      {block.description && !block.subItem && (
+        <Help title={block.description} />
+      )}
+    </div>
+  );
+};
