+++ packages/easy-email-extensions/src/KiaLayout/leftnav/index.tsx
@@ -1,0 +1,1 @@
+export * from './LeftPanel';
